Componentize with new component inpuit
